+++ src/prover/data_structures/top_symbol_hashing.rs
